♻️ refactor: rating 타입 double로 수정

+++ src/main/java/com/sparta/goodbite/domain/review/dto/CreateReviewRequestDto.java
@@ -9,9 +9,11 @@ public abstract class CreateReviewRequestDto {
 
     @NotNull(message = "평점을 입력해 주세요.")
     @RatingConstraint
-    protected float rating;
+    protected double rating;
+
     @NotNull(message = "리뷰 내용을 입력해 주세요.")
     protected String content;
+
     @NotNull(message = "식당 ID를 입력해 주세요.")
     private Long restaurantId;
 }
